+++ CHANGES.md
@@ -3,4 +3,5 @@ This is a fork of [DiffNav](https://github.com/dlvhdr/diffnav) with some extra f
 Changes with respect to upstream:
 - light theme support,
 - select & copy to clipboard in the diff view window,
-- OSC 52 clipboard instead of atotto/clipboard.
+- OSC 52 clipboard instead of atotto/clipboard,
+- line wrapping in delta unified diff view.